+++ packages/tsurlfilter/src/rules/declarative-converter/declarative-rule.ts
@@ -9,6 +9,7 @@
 import { z as zod } from 'zod';
 
 import { RequestType } from '../../request-type';
+import { HTTPMethod } from '../../modifiers/method-modifier';
 
 /**
  * https://developer.chrome.com/docs/extensions/reference/declarativeNetRequest/#type-DomainType
@@ -151,6 +152,26 @@ export enum RequestMethod {
 }
 
 /**
+ * tsurlfilter {@link HTTPMethod} without {@link HTTPMethod.TRACE}
+ * because it is not supported by {@link RequestMethod}.
+ */
+export type SupportedHttpMethod = Exclude<HTTPMethod, HTTPMethod.TRACE>;
+
+/**
+ * Map {@link HTTPMethod} to declarative {@link RequestMethod}.
+ */
+export const DECLARATIVE_REQUEST_METHOD_MAP: Record<SupportedHttpMethod, RequestMethod> = {
+    [HTTPMethod.GET]: RequestMethod.Get,
+    [HTTPMethod.POST]: RequestMethod.Post,
+    [HTTPMethod.PUT]: RequestMethod.Put,
+    [HTTPMethod.DELETE]: RequestMethod.Delete,
+    [HTTPMethod.PATCH]: RequestMethod.Patch,
+    [HTTPMethod.HEAD]: RequestMethod.Head,
+    [HTTPMethod.OPTIONS]: RequestMethod.Options,
+    [HTTPMethod.CONNECT]: RequestMethod.Connect,
+};
+
+/**
  * https://developer.chrome.com/docs/extensions/reference/declarativeNetRequest/#type-RuleCondition
  */
 
@@ -165,7 +186,7 @@ const RuleConditionValidator = zod.strictObject({
     isUrlFilterCaseSensitive: zod.boolean().optional(),
     regexFilter: zod.string().optional(),
     requestDomains: zod.string().array().optional(),
-    requestMethods: zod.string().array().optional(),
+    requestMethods: zod.nativeEnum(RequestMethod).array().optional(),
     /**
      * If none of the `excludedResourceTypes` and `resourceTypes` are specified,
      * all resource types except "main_frame" will be matched.